check for dead philosopher

+++ philo/inc/philo.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   philo.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hbenfadd <hbenfadd@student.42.fr>          +#+  +:+       +#+        */
+/*   By: hamza <hamza@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/05 12:27:15 by hbenfadd          #+#    #+#             */
-/*   Updated: 2023/04/07 15:07:10 by hbenfadd         ###   ########.fr       */
+/*   Updated: 2023/04/08 10:16:42 by hamza            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ typedef struct s_philosophere
 {
 	int				id;
 	size_t			start;
-	int				last_eat;
+	long			last_eat;
 	pthread_mutex_t	check_death;
 	pthread_mutex_t	*rf;
 	pthread_mutex_t	*lf;
